OCR_Book_Extractor/requirements.txt
pytesseract
opencv-python
pillow
rapidfuzz
requests
fastapi
uvicorn
